+++ Changelog.md
@@ -1,3 +1,9 @@
+## 11.1.10
+* **Breaking** flags.midi-qol.onUseMacro Custom ItemMacro will no longer rewrite the change value when the effect is created on an actor. Rather it will be done when the effect is evaluated.
+  - This means that effects with flags.onUseMacro ItemMacro will need to be reapplied, in patricular any transfer effects, to "undo" the rewrite of ItemMacro -> ItemMacro.origin
+  - Requires DAE update to 11.0.22
+* Added first pass of creating api definitions for Monaco Macro editor for MidiQOL.function auto complete. I've not done much testing so feel free to respond with issues.
+
 ## 11.1.9
 * Fix for error thrown when removing template => remove concentration
 * Put back option to not show dead/unconscious when hp reach 0